ui/streamlit/requirements.txt
streamlit
streamlit-audiorecorder
azure-cognitiveservices-speech
openai
azure.identity
